refactor: 게시글 상세 정보 리턴 시 댓글 3개만 리턴

+++ src/main/java/com/depromeet/buzz/comment/repository/CommentRepository.java
@@ -11,7 +11,7 @@ import java.util.Optional;
 public interface CommentRepository extends JpaRepository<Comment, Long>, CommentRepositoryCustom {
     Optional<Comment> findById(Long commentId);
 
-    long countAllByPostId(Long postId);
+    int countAllByPostId(Long postId);
 
     List<Comment> findAllByPostId(Long postId);
 }
